+++ cmd/parse/wasm/code.go
@@ -1,0 +1,65 @@
+package wasm
+
+import (
+	"fmt"
+
+	modulestypes "github.com/forbole/bdjuno/v3/modules/types"
+	"github.com/forbole/bdjuno/v3/types"
+
+	parsecmdtypes "github.com/forbole/juno/v3/cmd/parse/types"
+	"github.com/forbole/juno/v3/types/config"
+	"github.com/spf13/cobra"
+
+	"github.com/forbole/bdjuno/v3/database"
+)
+
+// codeCmd returns the Cobra command allowing to fix all things related to x/wasm contract
+func codeCmd(parseConfig *parsecmdtypes.Config) *cobra.Command {
+	return &cobra.Command{
+		Use:   "codes",
+		Short: "Parse the x/wasm codes and store them in the database",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			parseCtx, err := parsecmdtypes.GetParserContext(config.Cfg, parseConfig)
+			if err != nil {
+				return err
+			}
+
+			sources, err := modulestypes.BuildSources(config.Cfg.Node, parseCtx.EncodingConfig)
+			if err != nil {
+				return err
+			}
+
+			// Get the database
+			db := database.Cast(parseCtx.Database)
+
+			height, err := db.GetLastBlockHeight()
+			if err != nil {
+				return err
+			}
+
+			// Get contract code list with infos
+			codesInfosRes, err := sources.WasmSource.GetCodesInfos(height)
+			if err != nil {
+				return fmt.Errorf("error while getting info of codes: %s", err)
+			}
+
+			var wasmCodes = make([]types.WasmCode, len(codesInfosRes))
+			for i, c := range codesInfosRes {
+				// Get code binary
+				bi, err := sources.WasmSource.GetCodeBinary(c.CodeID, height)
+				if err != nil {
+					return fmt.Errorf("error while getting code binary: %s", err)
+				}
+
+				wasmCodes[i] = types.NewWasmCode(c.Creator, bi, &c.InstantiatePermission, c.CodeID, height)
+			}
+
+			err = db.SaveWasmCodes(wasmCodes)
+			if err != nil {
+				return fmt.Errorf("error while saving wasm codes: %s", err)
+			}
+
+			return nil
+		},
+	}
+}